+++ client/bindings/nodejs/types/outputBuilderOptions/aliasOutputOptions.ts
@@ -1,13 +1,14 @@
 import type { FeatureTypes } from '@iota/types';
+import type { HexEncodedString } from '../../lib';
 import type { IBasicOutputBuilderOptions } from './basicOutputOptions';
 
 /**
  * Options for building an Alias Output
  */
 export interface IAliasOutputBuilderOptions extends IBasicOutputBuilderOptions {
-    aliasId: string;
+    aliasId: HexEncodedString;
     stateIndex?: number;
-    stateMetadata?: Uint8Array;
+    stateMetadata?: HexEncodedString;
     foundryCounter?: number;
     immutableFeatures?: FeatureTypes[];
 }